planner-main/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.32.0
PyJWT==2.8.0
pydantic==2.12.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.2
pytest==8.4.2
pytest-cov==5.0.0
coverage==7.5.0
prometheus-client==0.20.0
pyyaml==6.0.3
openpyxl==3.1.2

